Move rif file to FileInterface
